Cluster a BANKSY embedding

Cluster the BANKSY embedding obtained from different combinations of parameters (`lambda`, `use_agf`, `npcs`). Multiple clustering methods are implemented, with the default being Leiden graph-based clustering. For this method, parallelization is implemented on non-Windows operating systems.

This function allows a grid search over multiple parameters. Parameters which admit vectors are:

`lambda`

`use_agf`

`k.neighbors`

`mclust.G`

`kmeans.centers`

ClusterBanksy(
  bank,
  lambda = 0.2,
  use_agf = FALSE,
  pca = TRUE,
  npcs = 20,
  method = c("leiden", "louvain", "mclust", "kmeans"),
  k.neighbors = 50,
  resolution = 1,
  leiden.iter = -1,
  num.cores = 1,
  mclust.G = NULL,
  kmeans.centers = NULL,
  kmeans.iter.max = 10,
  seed = 1000,
  M = NULL,
  verbose = TRUE,
  ...
)

Arguments

bank

BanksyObject

lambda

(numeric) weighting parameter (default: 0.2 for celltyping, 0.8 for domain finding)

use_agf

(logical) vector specifying whether to use the AGF for clustering

pca

(logical) runs clustering on PCs (TRUE) or BANKSY matrix (FALSE)

npcs

(integer) number of pcs to use for clustering (default: 20)

method

(character) one of leiden, louvain, mclust, kmeans

k.neighbors

(integer) parameter for constructing sNN (for louvain / leiden)

resolution

(numeric) parameter used for clustering (leiden)

leiden.iter

(integer) number of leiden iterations (leiden)

num.cores

(integer) number of parallel cores on unix / macOS platforms (leiden)

mclust.G

(integer) number of mixture components (mclust)

kmeans.centers

(integer) number of clusters (kmeans)

kmeans.iter.max

(integer) max number of iterations (kmeans)

seed

(integer) random seed for reproducibility

M

(integer) advanced usage. vector specifying the highest azimuthal Fourier harmonic to cluster with. If specified, overwrites the use_agf argument.

verbose

(logical) show progress bar

...

to pass to methods

Value

BanksyObject with cluster labels in meta.data

Examples

# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
bank <- NormalizeBanksy(bank)
bank <- ComputeBanksy(bank)
#> Computing neighbors...
#> Spatial mode is kNN median
#> Parameters: k_geom = 15
#> Done
#> Computing harmonic m = 0
#> Using 15 neighbors
#> Done
bank <- ScaleBanksy(bank)
bank <- RunBanksyPCA(bank, lambda = 0.3)
#> Running PCA for M=0 lambda=0.3
#> BANKSY matrix with own.expr, F0
#> Squared lambdas: 0.7, 0.3
set.seed(100)
bank <- ClusterBanksy(bank, lambda = 0.3, npcs = 20, k.neighbors = 50, resolution = 0.5)