Sweep gene-cell matrix

Sweep gene-cell matrix

SweepBanksy(
  gcm,
  locs,
  rowparam = "lam",
  rowvalues = c(0, 0.25, 0.5),
  colparam = "res",
  colvalues = c(0.5, 1),
  sigma = 3,
  lambda = 0.4,
  alpha = 0.05,
  pcs = 15,
  k_expr = 30,
  res = 0.5,
  kspatial = 1000,
  k_geom = 10,
  normalizeColumns = TRUE,
  normalizeColumnsTo = 100,
  minimumTotalCounts = 0,
  lieden_iters = -1,
  zScaleRows = TRUE,
  zScaleBeforeAveraging = FALSE,
  zScaleOwnAfterAveraging = TRUE,
  zScaleNbrAfterAveraging = TRUE,
  zScaleDatasetsIndividually = TRUE,
  multiple_datasets = FALSE,
  cells_split_by_dataset = NULL,
  dimensions = "all",
  spatialMode = c("kNN_r", "kNearestNeighbours", "kNN_unif", "kNN_rank", "rNN_gauss"),
  banksyMatrices = NULL,
  compute_own_and_nbr_umap = FALSE,
  compute_3D_umap = FALSE,
  instructions
)

Arguments

gcm
gene-cell matrix, rows are features, columns are objects

locs
cell locations with object locations: x and y (and possibly z) coordinates. x and y are necessary, and must have at least two different locations. If data is in a single plane, it must be in a single z plane. The value of this z plane variable may be provided, but it is not necessary. There needs to be a column labeled cell_ID, which contains strings that match the column names of gcm, in the same order as the columns of gcm. locs should be a data.table with rows and columns as follows: sdimx sdimy sdimz cell_ID

rowparam
rowparam

rowvalues
rowvals

colparam
colparam

colvalues
colvals

sigma
sigma

lambda
lambda

alpha
alpha

pcs
pcs

k_expr
k_expr

res
res

kspatial
kspatial

k_geom
k_geom

normalizeColumns
norm

normalizeColumnsTo
norm

minimumTotalCounts
min

lieden_iters
niters

zScaleRows
z

zScaleBeforeAveraging
z

zScaleOwnAfterAveraging
z

zScaleNbrAfterAveraging
z

zScaleDatasetsIndividually
z

multiple_datasets
m

cells_split_by_dataset
cells

dimensions
dims

spatialMode
spatial

banksyMatrices
bank

compute_own_and_nbr_umap
no

compute_3D_umap
no

instructions
instrs

Value

Sweep List