Human dorsolateral prefrontal cortex 10x Visium data (sample 151675)

This dataset comprises 10x Visium profiling of the human DLPFC (Maynard et al. (2021) Transcriptome-scale spatial gene expression in the human dorsolateral prefrontal cortex. doi.org: 10.1038/s41593-020-00787-0) for subject 3 sample 151675. Gene expression for 3,592 spots in 2 dimensions was count normalized and subset to 7,319 genes obtained by taking the union of the top 3,000 genes in each sample for subject 3.

data(dlpfc151675)

Format

A list with 3 entries:

expression

(matrix) gene expression matrix

locations

(data.frame) spot locations in 2D

Source

http://research.libd.org/spatialLIBD/